Modifier: split header and implementation
